requirements.txt
flask==2.3.3
flask-login==0.6.2
pandas==2.0.3
openpyxl==3.1.2
psycopg2-binary==2.9.9
werkzeug==2.3.7
python-dotenv==1.0.0